Turns out that 'uptime' is a decimal.

+++ lib/services/models/provenance_validator.dart
@@ -58,5 +58,5 @@ class ProvenanceValidator {
   final String status;
   final String? imgUrl;
   final String? hr24Change;
-  final int uptime;
+  final double uptime;
 }
